Drop fed1 test
